+++ runtime/devnet/src/pallets/mod.rs
@@ -12,3 +12,9 @@ pub use balances::*;
 
 pub mod transaction_payment;
 pub use transaction_payment::*;
+
+pub mod scheduler;
+pub use scheduler::*;
+
+pub mod preimage;
+pub use preimage::*;
